+++ src/test/java/com/poorlex/poorlex/PoorlexApplicationTest.java
@@ -2,11 +2,9 @@ package com.poorlex.poorlex;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
-@ActiveProfiles("test")
 @SpringBootTest
-class PoorlexApplicationTests {
+class PoorlexApplicationTest {
 
     @Test
     void contextLoads() {
