+++ src/test/java/com/poorlex/poorlex/alarm/battlealarm/service/BattleAlarmServiceTest.java
@@ -27,7 +27,7 @@ import com.poorlex.poorlex.member.domain.Oauth2RegistrationId;
 import com.poorlex.poorlex.participate.domain.BattleParticipant;
 import com.poorlex.poorlex.participate.domain.BattleParticipantRepository;
 import com.poorlex.poorlex.support.ReplaceUnderScoreTest;
-import com.poorlex.poorlex.support.UsingDataJpaTest;
+import com.poorlex.poorlex.support.db.UsingDataJpaTest;
 import com.poorlex.poorlex.voting.vote.domain.Vote;
 import com.poorlex.poorlex.voting.vote.domain.VoteAmount;
 import com.poorlex.poorlex.voting.vote.domain.VoteDuration;
