+++ src/test/java/com/poorlex/poorlex/battleinvititation/documentation/BattleInviteDocumentationTest.java
@@ -5,6 +5,7 @@ import static org.mockito.BDDMockito.doNothing;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -51,6 +52,7 @@ class BattleInviteDocumentationTest extends RestDocsDocumentationTest implements
                 .header(HttpHeaders.AUTHORIZATION, "Bearer {accessToken}")
                 .content(objectMapper.writeValueAsString(request))
                 .contentType(MediaType.APPLICATION_JSON)
+                .with(csrf())
         );
         result.andExpect(status().isCreated())
             .andDo(
@@ -80,6 +82,7 @@ class BattleInviteDocumentationTest extends RestDocsDocumentationTest implements
                 .header(HttpHeaders.AUTHORIZATION, "Bearer {accessToken}")
                 .content(objectMapper.writeValueAsString(request))
                 .contentType(MediaType.APPLICATION_JSON)
+                .with(csrf())
         );
         result.andExpect(status().isOk())
             .andDo(
@@ -110,6 +113,7 @@ class BattleInviteDocumentationTest extends RestDocsDocumentationTest implements
                 .header(HttpHeaders.AUTHORIZATION, "Bearer {accessToken}")
                 .content(objectMapper.writeValueAsString(request))
                 .contentType(MediaType.APPLICATION_JSON)
+                .with(csrf())
         );
         result.andExpect(status().isOk())
             .andDo(
