+++ src/test/java/com/poorlex/poorlex/voting/votingpaper/service/VotingPaperServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import com.poorlex.poorlex.participate.domain.BattleParticipantRepository;
 import com.poorlex.poorlex.support.ReplaceUnderScoreTest;
-import com.poorlex.poorlex.support.UsingDataJpaTest;
+import com.poorlex.poorlex.support.db.UsingDataJpaTest;
 import com.poorlex.poorlex.voting.vote.domain.Vote;
 import com.poorlex.poorlex.voting.vote.domain.VoteAmount;
 import com.poorlex.poorlex.voting.vote.domain.VoteDuration;
